+++ src/main/java/org/broadinstitute/hellbender/tools/copynumber/formats/records/SimpleCount.java
@@ -13,7 +13,7 @@ import org.broadinstitute.hellbender.utils.param.ParamUtils;
  *
  * @author Samuel Lee &lt;slee@broadinstitute.org&gt;
  */
-public class SimpleCount implements Locatable, Feature {
+public final class SimpleCount implements Locatable, Feature {
 
     private final SimpleInterval interval;
     private final int count;
